x86/ipipe: fix detection of unhandled faults over the head domain
Commit #de4561efd introduced a regression when checking for the
current domain after a trap was propagated to the co-kernel. Since the
trap handler may have switched the calling context to the root domain,
we must not rely on the domain information detected on entry
(i.e. ipipe_root_p vs root_entry).

At this chance, we stop caching the domain id on entry, since there is
no performance gain in doing so.